feat: add responsive design to header
